Move eex format logic to Phases.Format.
